+++ drivers/lora/native/sx126x/sx126x.c
@@ -202,56 +202,6 @@ static int sx126x_set_rf_frequency(const struct device *dev, uint32_t freq)
 	return sx126x_hal_write_cmd(dev, SX126X_CMD_SET_RF_FREQUENCY, buf, 4);
 }
 
-static int sx126x_set_pa_config(const struct device *dev, uint8_t pa_duty_cycle,
-				uint8_t hp_max, uint8_t device_sel, uint8_t pa_lut)
-{
-	uint8_t buf[4] = { pa_duty_cycle, hp_max, device_sel, pa_lut };
-
-	return sx126x_hal_write_cmd(dev, SX126X_CMD_SET_PA_CONFIG, buf, 4);
-}
-
-static int sx126x_configure_pa_and_tx_params(const struct device *dev,
-					     int8_t power, uint32_t frequency,
-					     uint8_t ramp_time)
-{
-	const struct sx126x_hal_config *config = dev->config;
-	uint8_t pa_duty_cycle;
-	int8_t tx_power;
-	int ret;
-
-	if (config->is_sx1261) {
-		/*
-		 * SX1261: Low power PA, up to +15 dBm
-		 * For +15 dBm at >400 MHz, use higher paDutyCycle.
-		 * For lower power, use lower paDutyCycle for efficiency.
-		 */
-		pa_duty_cycle = (power >= SX1261_MAX_POWER && frequency >= 400000000)
-				? SX1261_PA_DUTY_CYCLE_HIGH
-				: SX1261_PA_DUTY_CYCLE_LOW;
-		ret = sx126x_set_pa_config(dev, pa_duty_cycle, SX1261_HP_MAX,
-					   SX126X_DEVICE_SEL_SX1261,
-					   SX126X_PA_LUT);
-		if (ret < 0) {
-			return ret;
-		}
-		tx_power = CLAMP(power, SX1261_MIN_POWER, SX1261_MAX_POWER_TX_PARAM);
-	} else {
-		/* SX1262: High power PA, up to +22 dBm */
-		ret = sx126x_set_pa_config(dev, SX1262_PA_DUTY_CYCLE,
-					   SX1262_HP_MAX,
-					   SX126X_DEVICE_SEL_SX1262,
-					   SX126X_PA_LUT);
-		if (ret < 0) {
-			return ret;
-		}
-		tx_power = CLAMP(power, SX1262_MIN_POWER, SX1262_MAX_POWER);
-	}
-
-	uint8_t buf[2] = { (uint8_t)tx_power, ramp_time };
-
-	return sx126x_hal_write_cmd(dev, SX126X_CMD_SET_TX_PARAMS, buf, 2);
-}
-
 static int sx126x_set_modulation_params(const struct device *dev,
 					uint8_t sf, uint8_t bw, uint8_t cr,
 					bool ldro)
@@ -575,6 +525,9 @@ static void sx126x_irq_work_handler(struct k_work *work)
 	if (irq_status & SX126X_IRQ_RX_TX_TIMEOUT) {
 		sx126x_handle_irq_timeout(dev);
 	}
+
+	/* Re-enable the DIO1 interrupt for the next event */
+	sx126x_hal_dio1_irq_enable(dev);
 }
 
 static int sx126x_lora_config(const struct device *dev,
@@ -603,7 +556,7 @@ static int sx126x_lora_config(const struct device *dev,
 	}
 
 	/* Configure PA and TX power based on chip variant and frequency */
-	ret = sx126x_configure_pa_and_tx_params(dev, config->tx_power,
+	ret = sx126x_hal_configure_tx_params(dev, config->tx_power,
 						config->frequency,
 						SX126X_RAMP_200_US);
 	if (ret < 0) {
@@ -942,7 +895,7 @@ static int sx126x_lora_test_cw(const struct device *dev, uint32_t frequency,
 	}
 
 	/* Set PA config and TX power */
-	ret = sx126x_configure_pa_and_tx_params(dev, tx_power, frequency,
+	ret = sx126x_hal_configure_tx_params(dev, tx_power, frequency,
 						SX126X_RAMP_200_US);
 	if (ret < 0) {
 		k_mutex_unlock(&data->lock);
